chore: move test utils to test_utils.py
